+++ frontend/src/App.tsx
@@ -9,6 +9,7 @@ import { Inventory } from './pages/Inventory'
 import { HelmPage } from './pages/HelmPage'
 import { Nodes } from './pages/Nodes'
 import { Integrations } from './pages/Integrations'
+import { Secrets } from './pages/Secrets'
 import { ClusterDetail } from './pages/ClusterDetail'
 
 function ProtectedRoutes() {
@@ -40,6 +41,7 @@ function ProtectedRoutes() {
           <Route path="/inventory" element={<Inventory />} />
           <Route path="/helm" element={<HelmPage />} />
           <Route path="/nodes" element={<Nodes />} />
+          <Route path="/secrets" element={<Secrets />} />
           <Route path="/integrations" element={<Integrations />} />
           <Route path="/clusters/:id" element={<ClusterDetail />} />
           <Route path="*" element={<Navigate to="/" replace />} />